- SPEA2 is implemented - Tests are executed and graphics are updated. Results are added to journal v3.1.

+++ MSThesis-resurrected/MS Thesis/src/tr/edu/metu/ceng/ms/thesis/modstarlite/data/Objective.java
@@ -37,7 +37,7 @@ public class Objective {
 	}
 
 	public String toShortenString() {
-		return String.valueOf(this.getValue());
+		return String.valueOf((int)this.getValue());
 	}
 
 	private String getShortenedBehaviour() {
